requirements.txt
openai
langchain
langchain_ai21
langchain-community
streamlit
python-dotenv
PyPDF2
pandas
langchain-huggingface 
huggingface_hub

#for installation of local packages
-e .